apps/configurator/pyproject.toml
[project]
name = "configurator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.2",
    "pyyaml>=6.0.3",
    "structlog>=25.5.0",
    "uvicorn>=0.42.0",
]
